Get the ANU logo

A quick way of getting the logo. Not all combinations of category, type, and color (see table below).

primary	secondary	crest	favicon

vertical	Y	Y

horizontal	Y	Y

inversed			Y	Y

outline			Y

gold_black	Y	Y

gold_white	Y	Y

black	Y	Y	Y	Y

white	Y	Y	Y	Y

gold			Y	Y

Usage

logo_get(
  path = ".",
  overwrite = FALSE,
  color_mode = c("RGB", "CMYK"),
  category = c("primary", "secondary", "crest", "favicon"),
  type = if (category[1] %in% c("primary", "secondary")) c("vertical", "horizontal")
    else c("inversed", "outline"),
  color = if (category[1] %in% c("primary", "secondary")) c("gold_black", "gold_white",
    "black", "white") else c("black", "white", "gold"),
  ext = c("png", "jpg", "eps"),
  filename = NULL
)

Source

https://anu365.sharepoint.com/sites/ANUidentityHUB/SitePages/ANU-logo.aspx

Arguments

path

The path to save the logo to.

overwrite

Overwrite the logo if it already exists.

color_mode

Use RGB for digital communication (e.g. website) and CMYK for printing.

category

The category of the logo.

type

The type of the logo. Note that some type do not exist based on category chosen.

color

The color of the logo. Note that some colors do not exist based on category and type chosen.

ext

The file extension of the logo. Note that some extensions do not exist based on category, type and color chosen.

filename

The output filename of the logo (defaults to the original filename).